% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get.data}
\alias{get.data}
\title{Read data from a google sheet}
\usage{
get.data(url, skip = 1)
}
\arguments{
\item{url}{URL to the google workbook you want to read}

\item{skip}{The number of lines to skip before reading data from google sheets}
}
\value{
An r object with the data
}
\description{
Reads data from a sheets in a google sheet
}
