% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{read.all.sheets}
\alias{read.all.sheets}
\title{Read data from a google sheet}
\usage{
read.all.sheets(sheet, url, skip = 1)
}
\arguments{
\item{sheet}{The sheetname to read}

\item{url}{URL to the google workbook you want to read}

\item{skip}{The number of lines to skip before reading data, default to 1}
}
\value{
An r object with the data
}
\description{
Reads data from a sheets in a google sheet
}
